Move to using a hook, introduce heartbeat component
