+++ ui/components/app/edit-gas-popover/edit-gas-popover.component.js
@@ -1,6 +1,5 @@
-import React, { useCallback, useContext, useState, useRef } from 'react';
+import React, { useCallback, useContext, useState } from 'react';
 import PropTypes from 'prop-types';
-import classNames from 'classnames';
 
 import { useDispatch, useSelector } from 'react-redux';
 import { useGasFeeInputs } from '../../../hooks/useGasFeeInputs';
@@ -29,6 +28,7 @@ import {
   hideSidebar,
   updateTransaction,
 } from '../../../store/actions';
+import LoadingHeartBeat from '../../ui/loading-heartbeat';
 
 export default function EditGasPopover({
   popoverTitle = '',
@@ -43,8 +43,7 @@ export default function EditGasPopover({
   const dispatch = useDispatch();
   const showSidebar = useSelector((state) => state.appState.sidebar.isOpen);
 
-  const editDisplayNode = useRef(null);
-  const loadingAnimationClass = useShouldAnimateGasEstimations(editDisplayNode);
+  const shouldAnimate = useShouldAnimateGasEstimations();
 
   const showEducationButton =
     mode === EDIT_GAS_MODES.MODIFY_IN_PLACE && process.env.SHOW_EIP_1559_UI;
@@ -186,47 +185,48 @@ export default function EditGasPopover({
         )
       }
     >
-      <div style={{ padding: '0 20px 20px 20px' }}>
+      <div style={{ padding: '0 20px 20px 20px', position: 'relative' }}>
         {showEducationContent ? (
           <EditGasDisplayEducation />
         ) : (
-          <EditGasDisplay
-            ref={editDisplayNode}
-            className={classNames(loadingAnimationClass)}
-            showEducationButton={showEducationButton}
-            warning={warning}
-            showAdvancedForm={showAdvancedForm}
-            setShowAdvancedForm={setShowAdvancedForm}
-            dappSuggestedGasFeeAcknowledged={dappSuggestedGasFeeAcknowledged}
-            setDappSuggestedGasFeeAcknowledged={
-              setDappSuggestedGasFeeAcknowledged
-            }
-            maxPriorityFeePerGas={maxPriorityFeePerGas}
-            setMaxPriorityFeePerGas={setMaxPriorityFeePerGas}
-            maxPriorityFeePerGasFiat={maxPriorityFeePerGasFiat}
-            maxFeePerGas={maxFeePerGas}
-            setMaxFeePerGas={setMaxFeePerGas}
-            maxFeePerGasFiat={maxFeePerGasFiat}
-            estimatedMaximumNative={estimatedMaximumNative}
-            isGasEstimatesLoading={isGasEstimatesLoading}
-            gasFeeEstimates={gasFeeEstimates}
-            gasEstimateType={gasEstimateType}
-            gasPrice={gasPrice}
-            setGasPrice={setGasPrice}
-            gasLimit={gasLimit}
-            setGasLimit={setGasLimit}
-            estimateToUse={estimateToUse}
-            setEstimateToUse={setEstimateToUse}
-            estimatedMinimumFiat={estimatedMinimumFiat}
-            estimatedMaximumFiat={estimatedMaximumFiat}
-            hasGasErrors={hasGasErrors}
-            onEducationClick={() => setShowEducationContent(true)}
-            mode={mode}
-            transaction={transaction}
-            gasErrors={gasErrors}
-            onManualChange={onManualChange}
-            {...editGasDisplayProps}
-          />
+          <>
+            <LoadingHeartBeat active={shouldAnimate} />
+            <EditGasDisplay
+              showEducationButton={showEducationButton}
+              warning={warning}
+              showAdvancedForm={showAdvancedForm}
+              setShowAdvancedForm={setShowAdvancedForm}
+              dappSuggestedGasFeeAcknowledged={dappSuggestedGasFeeAcknowledged}
+              setDappSuggestedGasFeeAcknowledged={
+                setDappSuggestedGasFeeAcknowledged
+              }
+              maxPriorityFeePerGas={maxPriorityFeePerGas}
+              setMaxPriorityFeePerGas={setMaxPriorityFeePerGas}
+              maxPriorityFeePerGasFiat={maxPriorityFeePerGasFiat}
+              maxFeePerGas={maxFeePerGas}
+              setMaxFeePerGas={setMaxFeePerGas}
+              maxFeePerGasFiat={maxFeePerGasFiat}
+              estimatedMaximumNative={estimatedMaximumNative}
+              isGasEstimatesLoading={isGasEstimatesLoading}
+              gasFeeEstimates={gasFeeEstimates}
+              gasEstimateType={gasEstimateType}
+              gasPrice={gasPrice}
+              setGasPrice={setGasPrice}
+              gasLimit={gasLimit}
+              setGasLimit={setGasLimit}
+              estimateToUse={estimateToUse}
+              setEstimateToUse={setEstimateToUse}
+              estimatedMinimumFiat={estimatedMinimumFiat}
+              estimatedMaximumFiat={estimatedMaximumFiat}
+              onEducationClick={() => setShowEducationContent(true)}
+              mode={mode}
+              transaction={transaction}
+              hasGasErrors={hasGasErrors}
+              gasErrors={gasErrors}	
+              onManualChange={onManualChange}
+              {...editGasDisplayProps}
+            />
+          </>
         )}
       </div>
     </Popover>
