+++ ui/components/app/transaction-detail/index.scss
@@ -1,6 +1,5 @@
 .transaction-detail {
-  opacity: 1;
-  transition: opacity 0.5s ease-in-out;
+  position: relative;
 
   .transaction-detail-edit {
     text-align: end;
@@ -29,9 +28,5 @@
       padding-bottom: 0;
       border-bottom: 0;
     }
-  }
-
-  &--loading {
-    opacity: 0;
   }
 }
