+++ ui/components/app/transaction-detail/transaction-detail.component.js
@@ -1,23 +1,19 @@
-import React, { useContext, useRef } from 'react';
+import React, { useContext } from 'react';
 import PropTypes from 'prop-types';
-import classNames from 'classnames';
 
 import { I18nContext } from '../../../contexts/i18n';
 import { useShouldAnimateGasEstimations } from '../../../hooks/useShouldAnimateGasEstimations';
 
 import TransactionDetailItem from '../transaction-detail-item/transaction-detail-item.component';
+import LoadingHeartBeat from '../../ui/loading-heartbeat';
 
 export default function TransactionDetail({ rows = [], onEdit }) {
   const t = useContext(I18nContext);
-
-  const containerNode = useRef(null);
-  const loadingAnimationClass = useShouldAnimateGasEstimations(containerNode);
+  const shouldAnimate = useShouldAnimateGasEstimations();
 
   return (
-    <div
-      ref={containerNode}
-      className={classNames('transaction-detail', loadingAnimationClass)}
-    >
+    <div className="transaction-detail">
+      <LoadingHeartBeat active={shouldAnimate} />
       {onEdit && (
         <div className="transaction-detail-edit">
           <button onClick={onEdit}>{t('edit')}</button>
