+++ ui/components/ui/loading-heartbeat/index.scss
@@ -1,0 +1,23 @@
+.loading-heartbeat {
+  position: absolute;
+  top: 0;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  opacity: 0;
+  background: #fff;
+  display: none;
+
+  &--active {
+    display: block;
+    animation: heartbeat 2s ease-in-out;
+  }
+}
+
+@keyframes heartbeat {
+  0% { opacity: 0; }
+  25% { opacity: 1; }
+  50% { opacity: 0.5; }
+  75% { opacity: 1; }
+  100% { opacity: 0; }
+}
