+++ ui/components/ui/ui-components.scss
@@ -31,6 +31,7 @@
 @import 'identicon/index';
 @import 'info-tooltip/index';
 @import 'list-item/index';
+@import 'loading-heartbeat/index';
 @import 'loading-indicator/loading-indicator';
 @import 'loading-screen/index';
 @import 'menu/menu';
